pyproject.toml
[project]
name = "rag-prj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-core>=1.2.28",
    "langchain-openai>=1.1.12",
    "python-dotenv>=1.2.2",
]
